requirements.txt
# Pollen Feature Extraction - Dependencies
# Cài đặt: pip install -r requirements.txt

# Xử lý ảnh
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Machine Learning & Deep Learning
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Database
sqlite3-api>=2.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Image augmentation
albumentations>=1.3.0

# Feature extraction
mahotas>=1.4.0
